Update Dynamic or late binding
Run this example in your IDE and pay attention to the invoked display function in the Super_Class. I have learned that the binding between the origin of the virtual function (inside the Super_Class) and its replacement (defined within the subclasses) is created dynamically, during the execution of the program. This is also known as late binding and occurs and runtime.